+++ frontend/src/panel/scripting_interface/user_scripts/UnpublishedScriptListItem.tsx
@@ -6,7 +6,7 @@ import CardContent from '@mui/material/CardContent/CardContent';
 import Typography from '@mui/material/Typography/Typography';
 import CardActions from '@mui/material/CardActions/CardActions';
 import { editorScreen } from '../../navigation/screens';
-import { UnpublishedScriptWithWebsite } from '../../models/api_temp/UnpublishedScript';
+import { UnpublishedScriptWithWebsite } from '../../models/api/UnpublishedScript';
 
 const UnpublishedScriptListItem: React.FC<{
   script: UnpublishedScriptWithWebsite;
