+++ frontend/src/panel/support_interface/script_support/ScriptSupport.tsx
@@ -5,7 +5,7 @@ import SupportDialog from './components/SupportDialog';
 import SupportButton from './components/SupportButton';
 import InstructionsDisplay from './components/InstructionsDisplay';
 import useScriptSupport from './useScriptSupport';
-import { Script } from '../../models/api_temp/Script';
+import { Script } from '../../models/api/Script';
 
 interface ScriptSupportProps {
   script: Script;
