+++ frontend/src/panel/support_interface/script_support/components/ScriptDetails.tsx
@@ -1,5 +1,5 @@
 import Stack from '@mui/material/Stack/Stack';
-import { Script } from '../../../models/api_temp/Script';
+import { Script } from '../../../models/api/Script';
 import Typography from '@mui/material/Typography/Typography';
 
 interface ScriptDetailsProps {
